+++ types/decimal-to-binary.h
@@ -1,6 +1,6 @@
 #ifndef DECIMAL_TO_BINARY_H
 #define DECIMAL_TO_BINARY_H
 
-int decimalToBinary(unsigned long long n, int binary[]);
+int decimalToBinary(int n, int binary[]);
 
 #endif
