+++ types/modular-exponentiation.h
@@ -1,6 +1,6 @@
 #ifndef MODULAR_EXPONENTIATION_H
 #define MODULAR_EXPONENTIATION_H
 
-unsigned long long modularExponentiation(unsigned long long base, unsigned long long mod, int binaryExp[], int size);
+int modularExponentiation(int base, int mod, int binaryExp[], int size);
 
 #endif
